[Fix] PluginBridge: Even though InterlockedSomething WinAPI functions work with uint32_t with the VS2010 provided Windows SDK, this is not documented. Older SDK versions do not provide unsigned overloads for C++ and thus compilation fails with VS2008. Replace uint32_t with the documented supported type for the InterlockedSomething functions: LONG.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3914 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ pluginBridge/Bridge.cpp
@@ -43,7 +43,7 @@
 #include "../common/thread.h"
 
 
-uint32_t PluginBridge::instanceCount = 0;
+LONG PluginBridge::instanceCount = 0;
 Event PluginBridge::sigQuit;
 
 // This is kind of a back-up pointer in case we couldn't sneak our pointer into the AEffect struct yet.
